requirements.txt
absl-py==0.10.0
aioredis==1.3.1
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
av==8.0.2
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.3.0
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.1
dash==1.17.0
dash-bootstrap-components==0.10.7
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
decorator==4.4.2
defusedxml==0.7.0rc1
dj-pagination==2.5.0
Django==3.1.7
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-cacheops==5.0.1
django-compressor==2.4
django-filter==2.3.0
django-mysql==3.3.0
django-plotly-dash==1.4.2
django-redis==4.12.1
django-rest-auth==0.9.5
django-revproxy==0.10.0
django-rq==2.3.2
django-sendfile==0.3.11
django-treebeard==4.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dpd-components==0.1.0
dpd-static-support==0.0.5
drf-yasg==1.17.1
EasyProcess==0.3
entrypoint2==0.2.1
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.8.0
flatbuffers==1.12
funcy==1.14
furl==2.1.0
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.28.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hiredis==1.1.0
httplib2==0.19.0
hyperlink==20.0.1
idna==2.9
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.0
mysqlclient==1.4.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.5.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-logstash==0.4.6
python3-openid==3.1.0
pytz==2019.3
pyunpack==0.2.1
PyYAML==5.4.1
pyzmq==19.0.2
rcssmin==1.0.6
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rjsmin==1.1.0
rq==1.5.2
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
rules==2.2
scipy==1.4.1
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
six==1.15.0
smmap==3.0.4
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.2.0